(hotfix): resolve incorrect and inconsistent module naming
- see https://github.com/agilgur5/django-serializable-model/issues/2
- in Python-land, packages are _not_ imported by their package name,
  but their _module_ name
  - see also https://stackoverflow.com/a/54599368/3431180
  - and the module name of this package was `serializable`, _not_
    `django-serializable-model`
- also in Python-land, one cannot use the standard `import` syntax
  for hyphenated modules, it will actually cause a SyntaxError
  - but for some reason it's not illegal to do so, and is allowed both
    in `setuptools` naming and PyPI naming (it is discouraged by PEP8)

- rename serializable.py to django_serializable_model.py
  - leave serializable.py and just make it a wrapper / alias of
    django_serialiazable_model so as not to cause a breaking change

(pkg): publish `django_serializable_model` module as well
- don't remove `serializable` though so as to not cause a breaking
  change; should remove it in v1.0.0

(docs): use `django_serializable_model` in Usage docs
